refactor: move condition code
