+++ package.json
@@ -7,8 +7,8 @@
     "dev": "node scripts/dev.js",
     "build": "parcel build packages/main/dist.mjs -d dist --out-file xhear.js --no-source-maps --global global",
     "static-server": "node scripts/static-server",
-    "test": "npm run build && npx playwright test",
-    "ci-test": "node scripts/ci-test"
+    "playwright": "npm run build && npx playwright test",
+    "test": "node scripts/ci-test"
   },
   "author": {
     "name": "YAO",
